refactor: move logic from navbar to parent
