+++ src/components/TheNav.tsx
@@ -1,27 +1,15 @@
-import { NavLink, useNavigate } from 'react-router-dom'
+import { NavLink } from 'react-router-dom'
 import { Clock8Icon, HeartIcon, LogOutIcon } from 'lucide-react'
-import { toast } from 'sonner'
+import type { User } from '@supabase/supabase-js'
 import { Button } from '@/components/ui/button'
-import { useAppDispatch, useAppSelector } from '@/hooks/reduxHooks'
-import supabase from '@/services/supabase'
-import { setUser } from '@/redux/slices/authSlice'
 
 interface Props {
   children?: React.ReactNode
+  handleSignOut: () => void
+  user: User | null
 }
 
-export function TheNav({ children }: Props) {
-  const user = useAppSelector(state => state.auth.user)
-  const dispatch = useAppDispatch()
-  const navigate = useNavigate()
-
-  async function signOut() {
-    await supabase.auth.signOut()
-    dispatch(setUser(false))
-    navigate('/login')
-    toast.success('signed out!')
-  }
-
+export function TheNav({ children, handleSignOut, user }: Props) {
   return (
     <nav className="flex items-center gap-2">
       {children}
@@ -38,7 +26,7 @@ export function TheNav({ children }: Props) {
                 <Clock8Icon className="h-[1.2rem] w-[1.2rem] scale-100 transition-all " />
               </NavLink>
             </Button>
-            <Button onClick={signOut} variant="outline" size="icon">
+            <Button onClick={handleSignOut} variant="outline" size="icon">
               <LogOutIcon className="h-[1.2rem] w-[1.2rem] scale-100 transition-all " />
             </Button>
